requirements.txt
google-generativeai
python-dotenv
psutil
pycryptodome
requests
colorama